build error gone

+++ pwm.h
@@ -5,13 +5,14 @@
 #include "timer.h"
 #include <xc.h>
 #include "stdint.h"
+#include "setup.h"
 
 #define _XTAL_FREQ 12000000 //12MHz
 #define MAX_CANARD_ANGLE 10 //+- 10 degrees
 
 void pwm_init(void);
 
-void updatePulseWidth(uint8_t angle);
+void updatePulseWidth(uint16_t angle);
 
 #endif	/* PWM_H */
